requirements.txt
asgiref>=3.7.2
django>=4.2.11
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
mysqlclient>=2.2.4
pillow>=10.2.0
pip>=24.0
pytz>=2024.1
setuptools>=58.0.4
sqlparse>=0.4.4
typing_extensions>=4.10.0
